py_backend/requirements.txt
fastapi>=0.98.0
uvicorn>=0.22.0
boto3>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pymongo>=4.4.0
mangum>=0.17.0  # For AWS Lambda deployment (optional)
pytest>=7.3.1  # For testing
httpx>=0.24.1  # For testing
